Career-Advisor-Chatbot/requirements.txt
streamlit>=1.31.0
google-generativeai>=0.4.0
python-dotenv>=1.0.0
streamlit-authenticator==0.3.2